+++ Code/Views/LYRUIMessageInputToolbar.h
@@ -62,8 +62,8 @@ extern NSString *const LYRUIMessageInputToolbarDidChangeHeightNotification;
 /**
  @abstract The `LYRUIMessageInputToolbar` provides a lightweight and customizable class
  that is similar to the message input toolbar in Messages.
- @discussion The class displays two customizable UIButtons seperated by a message input text view.
- The class automatically resizes itself in response to user or system input content. The view also
+ @discussion The class displays two customizable `UIButton` objects separated by a message input text view.
+ Instances are automatically resized in response to user or system input. The view also
  caches any content provided and exposes that content back to a consuming object via the
  messageParts property.
  */